+++ docs/examples/03_backends_ros/files/02_robot_model.py
@@ -4,4 +4,4 @@ with RosClient() as client:
     robot = client.load_robot()
     robot.info()
 
-    assert robot.name == 'ur5'
+    assert robot.name == 'ur5_robot'