+++ docs/examples/03_backends_ros/files/02_robot_model_urdf.py
@@ -1,14 +1,12 @@
 from compas.robots import RobotModel
 
-import compas_fab
 from compas_fab.backends import RosClient
 from compas_fab.robots import Robot
 
 with RosClient() as client:
-    urdf = compas_fab.get('universal_robot/ur_description/urdf/ur5.urdf')
-    model = RobotModel.from_urdf_file(urdf)
+    model = RobotModel.ur5()
     robot = Robot(model, client=client)
 
     robot.info()
 
-    assert robot.name == 'ur5'
+    assert len(robot.get_configurable_joint_names()) == 6