+++ docs/examples/03_backends_ros/files/04_plan_motion.py
@@ -1,17 +1,19 @@
 import math
+
 from compas.geometry import Frame
-from compas.robots import Configuration
+
 from compas_fab.backends import RosClient
 
 with RosClient() as client:
     robot = client.load_robot()
-    assert robot.name == 'ur5'
+    assert robot.name == 'ur5_robot'
 
     frame = Frame([0.4, 0.3, 0.4], [0, 1, 0], [0, 0, 1])
     tolerance_position = 0.001
     tolerance_axes = [math.radians(1)] * 3
 
-    start_configuration = Configuration.from_revolute_values([-0.042, 4.295, 0, -3.327, 4.755, 0.])
+    start_configuration = robot.zero_configuration()
+    start_configuration.joint_values = (-3.530, 3.830, -0.580, -3.330, 4.760, 0.000)
     group = robot.main_group_name
 
     # create goal constraints from frame
@@ -24,7 +26,7 @@ with RosClient() as client:
                                    start_configuration,
                                    group,
                                    options=dict(
-                                       planner_id='RRTstarkConfigDefault'
+                                       planner_id='RRTConnect'
                                    ))
 
     print("Computed kinematic path with %d configurations." % len(trajectory.points))