+++ src/compas_fab/backends/ros/backend_features/move_it_plan_motion.py
@@ -2,25 +2,21 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import itertools
-
-from compas.robots import Configuration
 from compas.utilities import await_callback
 
 from compas_fab.backends.interfaces import PlanMotion
 from compas_fab.backends.ros.backend_features.helpers import convert_constraints_to_rosmsg
-from compas_fab.backends.ros.backend_features.helpers import convert_trajectory_points
+from compas_fab.backends.ros.backend_features.helpers import convert_trajectory
 from compas_fab.backends.ros.backend_features.helpers import validate_response
 from compas_fab.backends.ros.messages import AttachedCollisionObject
 from compas_fab.backends.ros.messages import Header
 from compas_fab.backends.ros.messages import JointState
-from compas_fab.backends.ros.messages import MotionPlanResponse
 from compas_fab.backends.ros.messages import MotionPlanRequest
+from compas_fab.backends.ros.messages import MotionPlanResponse
 from compas_fab.backends.ros.messages import MultiDOFJointState
 from compas_fab.backends.ros.messages import RobotState
 from compas_fab.backends.ros.messages import TrajectoryConstraints
 from compas_fab.backends.ros.service_description import ServiceDescription
-from compas_fab.robots import JointTrajectory
 
 __all__ = [
     'MoveItPlanMotion'
@@ -68,7 +64,7 @@ class MoveItPlanMotion(PlanMotion):
               violates these constraints. Defaults to ``None``.
             - ``"planner_id"``: (:obj:`str`)
               The name of the algorithm used for path planning.
-              Defaults to ``'RRTConnectkConfigDefault'``.
+              Defaults to ``'RRTConnect'``.
             - ``"num_planning_attempts"``: (:obj:`int`, optional)
               Normally, if one motion plan is needed, one motion plan is computed.
               However, for algorithms that use randomization in their execution
@@ -100,21 +96,17 @@ class MoveItPlanMotion(PlanMotion):
 
         # Use base_link or fallback to model's root link
         options['base_link'] = options.get('base_link', robot.model.root.name)
-        options['joint_names'] = robot.get_configurable_joint_names()
-        options['joint_types'] = robot.get_configurable_joint_types()
+        options['joints'] = {j.name: j.type for j in robot.model.joints}
 
         return await_callback(self.plan_motion_async, **kwargs)
 
     def plan_motion_async(self, callback, errback,
                           goal_constraints, start_configuration=None, group=None, options=None):
         """Asynchronous handler of MoveIt motion planner service."""
-
         # http://docs.ros.org/jade/api/moveit_core/html/utils_8cpp_source.html
         # TODO: if list of frames (goals) => receive multiple solutions?
-        joint_names = options['joint_names']
-        joint_types = options['joint_types']
-        joint_type_by_name = dict(zip(joint_names, joint_types))  # !!! should this go somewhere else? does it already exist somewhere else?
 
+        joints = options['joints']
         header = Header(frame_id=options['base_link'])
         joint_state = JointState(
             header=header,
@@ -140,7 +132,7 @@ class MoveItPlanMotion(PlanMotion):
                        goal_constraints=goal_constraints,
                        path_constraints=path_constraints,
                        trajectory_constraints=trajectory_constraints,
-                       planner_id=options.get('planner_id', 'RRTConnectkConfigDefault'),
+                       planner_id=options.get('planner_id', 'RRTConnect'),
                        group_name=group,
                        num_planning_attempts=options.get('num_planning_attempts', 1),
                        allowed_planning_time=options.get('allowed_planning_time', 2.),
@@ -148,24 +140,16 @@ class MoveItPlanMotion(PlanMotion):
                        max_acceleration_scaling_factor=options.get('max_acceleration_scaling_factor', 1.))
         # workspace_parameters=options.get('workspace_parameters')
 
-        def convert_to_trajectory(response):
-            trajectory = JointTrajectory()
-            trajectory.source_message = response
-            trajectory.fraction = 1.
-            trajectory.joint_names = response.trajectory.joint_trajectory.joint_names
-            trajectory.planning_time = response.planning_time
+        def response_handler(response):
+            try:
+                trajectory = convert_trajectory(joints,
+                                                response.trajectory,
+                                                response.trajectory_start,
+                                                1.,
+                                                response.planning_time,
+                                                response)
+                callback(trajectory)
+            except Exception as e:
+                errback(e)
 
-            joint_types = [joint_type_by_name[name] for name in trajectory.joint_names]
-            trajectory.points = convert_trajectory_points(
-                response.trajectory.joint_trajectory.points, joint_types)
-
-            start_state = response.trajectory_start.joint_state
-            start_state_types = [joint_type_by_name[name] for name in start_state.name]
-            trajectory.start_configuration = Configuration(start_state.position, start_state_types, start_state.name)
-            trajectory.attached_collision_meshes = list(itertools.chain(*[
-                aco.to_attached_collision_meshes()
-                for aco in response.trajectory_start.attached_collision_objects]))
-
-            callback(trajectory)
-
-        self.GET_MOTION_PLAN(self.ros_client, request, convert_to_trajectory, errback)
+        self.GET_MOTION_PLAN(self.ros_client, request, response_handler, errback)